ventas004

+++ src/main/java/com/back/interfaces/Identificable.java
@@ -1,6 +1,6 @@
 package com.back.interfaces;
 
 public interface  Identificable {
-    int getId();
-    void setId(int id);
+	Long getId();
+    void setId(Long id);
 }
